Math Editor: fix resizing issue
The math editor would not always be wide enough to show the whole
equation. This turned out to be because we were artificially limiting
its width when we sized its container from the space.

This commit removes this sizing so that the overlay width/height are
unconstrained and affect the green space's width/height (but not
vice versa), while the green space's left/top continue to position
the red overlay.

Test plan: typed long and tall equations in the box, saw the green
box always maintain space outside the equation :)